+++ eq_google_shopping_feed/__init__.py
@@ -21,3 +21,4 @@
 
 import controllers
 import models
+import google_product_taxonomy